Add files via upload
The script monitoring the status of multiple computers on a network. Using the PowerShell scripting language, the script is able to send pings to each computer and check for a response, providing a clear and concise overview of which machines are currently online and which are offline. The script is easy to use, and can be customized to suit the specific needs of any network administrator.
The script is GUI and all you need to do is change the name of the computers in the code.